Update zad2.cpp
Big bruh moment

+++ Pract. 01/zad2.cpp
@@ -3,8 +3,8 @@
 using namespace std;
 
 int main() {
-    int a = 5.4;
-    int b = 7.9;
+    double a = 5.4;
+    double b = 7.9;
 
     cout << "P = " << 2 * a + 2 * b << endl;
     cout << "S = " << a * b << endl;
